# s3_client: Increase default part size for optimal performance