@kreuzerkrieg 2025-02-06
> > client: Increase default part size for optimal performance #22732
> 
> Please use something like s3 client, for most developers client is the database's user.

Done